Foundations for a golang website.
The foundations for a golang website are created. The aboutme page
will hold information about me, and the Pages that are loaded from
the info handler are all pages that will be used for holding
various odd information that I want to display on my website that
is purely informational. The page is loaded from a template based
on the last section of the URI after /info/.

There is no main page that is currently designated because the root
handler was originally just testing the most simple type of website,
or just plain text. This will change later, and one of the
information pages will most likely become the root page.

The video page is just a page with a html5 video element that is
looking for a live stream. This feature is not finished yet, and
the server that will be serving up the video stream is not finished
either.

Assets are all served through the asset handler. They are
differentiated by the directory that is placed after the beginnning
/assets portion of the URI. For example a css page would be
/assets/css/page.css